refactor: detach ButtonContent struct #24
